Extract node-labeller.go from infra_test.go
Move helper method to libinfra
