+++ tests/infrastructure/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "crds.go",
         "downward-metrics.go",
         "k8s-client-changes.go",
+        "node-labeller.go",
         "prometheus.go",
         "taints-and-tolerations.go",
         "tls-configuration.go",
@@ -19,6 +20,7 @@ go_library(
         "//pkg/util/cluster:go_default_library",
         "//pkg/util/tls:go_default_library",
         "//pkg/virt-config:go_default_library",
+        "//pkg/virt-handler/node-labeller/util:go_default_library",
         "//pkg/virt-operator/resource/generate/components:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
@@ -26,6 +28,7 @@ go_library(
         "//tests/console:go_default_library",
         "//tests/containerdisk:go_default_library",
         "//tests/decorators:go_default_library",
+        "//tests/events:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/framework/checks:go_default_library",
