+++ tests/libinfra/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "leader.go",
         "metrics.go",
         "network.go",
+        "node-labeller.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/libinfra",
     visibility = ["//visibility:public"],
@@ -17,6 +18,7 @@ go_library(
         "//pkg/virt-controller/leaderelectionconfig:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
+        "//tests:go_default_library",
         "//tests/console:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
